feat/#131: 새로운 카테고리 API 반영

+++ Poppool/DataLayer/Data/Data/Network/API/CategoryAPI/ResponseDTO/GetCategoryListResponseDTO.swift
@@ -9,7 +9,7 @@ struct GetCategoryListResponseDTO: Codable {
 
 // MARK: - InterestResponse
 struct CategoryResponseDTO: Codable {
-    let categoryId: Int64
+    let categoryId: Int32
     let categoryName: String
 }
