+++ Poppool/DomainLayer/Domain/DomainInterface/Entity/AuthResponse/CategoryResponse.swift
@@ -1,11 +1,11 @@
 import Foundation
 
 public struct CategoryResponse {
-    public init(categoryId: Int64, category: String) {
-        self.categoryId = categoryId
+    public init(categoryId: Int32, category: String) {
+        self.categoryId = Int(categoryId)
         self.category = category
     }
 
-    public let categoryId: Int64
+    public let categoryId: Int
     public let category: String
 }
